+++ svc/pkg/linode/src/util/api.rs
@@ -502,11 +502,11 @@ pub struct ListInstanceTypesResponse {
 #[derive(Deserialize)]
 pub struct InstanceType {
 	pub id: String,
-	pub memory: u64,
-	pub disk: u64,
-	pub vcpus: u64,
-	pub transfer: u64,
-	pub network_out: u64,
+	pub memory: u32,
+	pub disk: u32,
+	pub vcpus: u32,
+	pub transfer: u32,
+	pub network_out: u32,
 }
 
 impl From<InstanceType> for crate::types::InstanceType {
